Added partial rendering throught javascript. The goal is to be able to create through the admin panel any type of partial and include them in any of your post.
The include pattern is to use <link rel='partial'
href='/my-partial-name' />

A small javascript snippet looks for those link and asynchronously load
them. Once loaded, it replaces the <link> tag with the actual partial.
This has the advantage that you can position your partial in your post
and the position will be kept.

This first commit has all the asynchronous logic but for now the
partials are coded in the project and cannot be modified in the admin
panel.