Admin menu overhaul
I've rebuild the whole admin menu to use a new algorithm I've been using
in La Marmite. I think it gives a bit of balance between convention and
customization.

On the other hand, I feel the callbacks used are a bit cumbersome. Some
of them could be actually remove so it's easier to implement custom
state without bring the 7 hells of the callbacks into play. I guess it
will be good for the time being but I'll keep a close eye on it. If I
follow up with my idea of building a gem out of this concept, I'll fix
it there.

The text is a bit blurry on antialiasing but I think it's a nice
incremental change from what it was before.