Added a meta description tag to help search engine.
Because I lack that meta before, it was showing arbitrary data in the
search result. Now it will show this sentence instead. This is a first
draft. I expect to change it in the near future as I get better ideas of
what I want to promote with this tag.