Fixed issue where I couldn't publish a post.
When refactoring away from engines, I forgot to change some manual form
input that were still using the old scheme. So, instead of using the
admin_post container, it was using post which would lead the controller
to think the user only want to update the post instead of publishing it.