Preview scrolling is now synchronized with the textarea
Fix #8 but I don't believe this is a final solution as the offset is
currently done at the top level and it will probably break when dealing
with the sub tabs (partials, images, etc). Also the CSS is a bit
repetitive on the form so it needs cleanup.

Nonetheless, the scrolling is now synchronized in a way that when you
edit, what you type will always stay on screen. Some issue I had when
opening a html tag and having the preview jump around is now resolved
with this implementation.