Modularized CSS and stabilizing UI
admin/sidebar.css should now hold everything that modifies the sidebar
abstractly. What I mean by that is that it should give a universal look
and feel in the admin panel. If a specific controller need to override
some specific look, it can do so within its own css file without
sidebar.css becoming a spaghetti. Also used a bit of mixin to make it
easier to understand the flow and break of indentation nightmares.

The new button is now part of a small slogan using the <main> part of
the screen in posts since I have yet to use that space and wanted to
give some balance to the screen.