Fixed the window glitch when the left menu grows.
CSS's height is a fixed value that can't grow. Because I'm using css
flow box, I expect the height value to grow as there's more item or the
screen shrink.

By changing height to min-height, I keep the base styling while keeping
responsiveness to the layout as the posts count grows or the screen
shrinks.