Image uploading is now A-W-E-S-O-M-E
Huge ass update that covers a bit too much but hey, it's alpha so I
guess I can deal with overlapping commits.

Uploads now works in the following way:

- Add S3's keys to config.local.yml
- In the admin panel, you open the subtab 'images'
- You can drag and drop an image and it will upload it directly to S3**
- Once your file is uploaded a thumbnails appear, you can drag and drop
  it to the location you want in your text area and the engine will
  paste the actual HTML tag preconfigured and ready to go.
- Done!

Seriously, I think I will upload images just for the hell of it because
it's so awesome.

**: Well not *directly*, you upload it to the server and the server
upload it to S3. I know I could upload directly to S3 which would save
bandwith but since uploading is not done heavily in blogs it was
unecessary. Moreover, the form upload would have required a callback
from S3 to handle the response which would have added complexity with
the validations/error handlings)