Removed JS exceptions. Fixed Partials system
So now the partials behave like the images. There's some glitches going
on with the loaded partials as it has to refetch the partial (even if
it's a 304, it will refetch from cache). So this is something that might
need tweaking for partials and images. Don't know yet so I'm leaving
this without issue.

Partials now have an overlay that gives user a feeling that drag is
possible.