Need an update view for when post update fails
This is similar to create.html.erb where it's only called when there's
an error. Maybe I should rename both to error.html and return them but
it gives an extra condition in the controller which can become
cumbersome.

Maybe a after callback could check for the record and render error...