Fixed the Javascript interpreter as first step to solve #15
Right now, I rebuild the whole scrip ton every input change. One of the
problem with doing that is the load that is incremental. One of the
solution would be to purge the DOM so every event bound to the DOM would
be flushed. This would probably give a flicker on the preview.

Since this is programming, I'm not sure there's a best way to do that.