Added dummy file in shared JS.
The reason is git doesn't allow to commit empty directory. I needed to
add something here as I want the shared/ folder to work and I needed to
resolve #26.

This small welcome has the advangtage of showing the user it can share
code between the blog and the post.

This fix #26.