Changed structure of controller to use a lazy loader
This also means that the post should be accessible as a method, not a
variable. I find it's easier to maintain a lazy loader as the post is
only set from one place. If another place needs a different type of
post, it might be a hint that a new controller is needed