Last big changes commit, explanation below
First, I reintroduced Warden's configuration to environment.rb

Then, I removed all the dependency in the gemspec as it slows down the
boot sequence too much for it to be any usable when developing on
Ecrire.

Most important iteration is I18n working now. I added a recursive path
so raisl will look at all the files in locales.