Removed set_application_directory!
Right now it's not useful as the binary is from the gem, it's not really
possible right now to figure it out like rails does. It may be means
that I'm wrong with this approach. Still don't know yet.

Another possibility would be to use config.ru.

Basically, it would be to move up the file hierarchy until the app meets
a config.ru file

This fix #39