Renamed path from theme:* to user:* and also monkey patched AR initializer
I renamed the path because it used to be there to create multiple
themes. Right now, only 1 theme is supported at any given time. Also,
since the railtie is now a descendent of Ecrire, it made more sense that
way.

I also monkey patched AR::Railtie.initializers to not crash when a
database cannot be initiated. Instead, if there's no database, the
system theme is used to let the user know he needs to configure his
database.

This fix #38