Well, I usually hate those huge ass long no meaning commits.
But hey, one thing led to another, you know. So here's a list of what
happens here.

When loading ecrire for the first time, it checks if the database is
configured and also if a User has been created. If not, it will load the
onboarding process so the user can configure Ecrire. Once it's finished
the user is asked to restart the server and the theme will be loaded.

Also modified the way Ecrire::Railtie loads the application, specially
the paths because it was loaded before it was possible to check if the
user existed.

Modified User model to use the same logic as has_secure_password. Needed
backward compatibility for now.

Theme is crap for now as I moved most of it out to onboarding.