Ecrire now looks recursively for config.ru to figure out blog's root.
Before, it was looking naively at Dir.pwd but this led to problem if you
are in the assets folder for example.

To fix this in a predictive manner, ecrire looks up the tree and try to
find config.ru if it finds it, it uses the directory path as the root
path for the blog.

If it can't find any, it raise an error that tells ecrire couldn't find
config.ru. It also suggests to create a new blog to the user.