Removed secrets template and dynamically generate it.
Since Ecrire needs to generate the secret_key_base that yaml file needed
to be parse & updated. It would have made sense to keep a template if
more information was in secrets.yml but it's not the case right now. So
what I did was build the secrets dynamically from the get go as it would
save code/manipulations.

I also refactored the 'new' commands as it was a placeholder (actually
it was the assets precompilation task...)

This fix #44