Simplified the onboarding process with the database
Before, the steps included 'variables' that the user had to change. Now
the controller generate dynamic values that are at the class level so as
long as the server is not restarted, the value stay the same. So the
user can try to connect many times and the value stay the same.