Tweaked the gemspec so compass actually works.
This commit makes me thinks that there's a problem with how I handle
dependencies. Because they are hardcoded, it becomes hard to adapt.

Bundler is the solution rails uses and I believe I should use the same.
I could use groups to specify what a user could modify (as the Gemfile
would live in the template's)