Removed database.yml from the template
database.yml should not be tracked, this way, you can have different
setup on different machines. The tracking was making things hard for me
to have a dev machine & prod machine with different settings.

Also, ecrire now checks if the database.yml file exist and recover if it
doesn't which wasn't the case before