Added placeholder to Admin::Post and modified the way Ecrire handles model to URL
Admin::Post would crash the editor if stylesheet, content or javascript
was set to nil. This problem could have been solved by using proper
default at the database level. However, this solution doesn't require
migration which I prefer for now as the migration system isn't top
notch.

As for the URL handling, posts cannot use their slug as URL as it's
possible a user will change that title through the editor. So in the
administration, the ID will be used and once it's published, the slug is
frozen which means we can use the slug as expected.