New editor that is just too hot to touch with your barehands.
Ok this commit wraps all the work that was necessary to build the new
editor. The editor as a simpler interface with autosave abilities as
well as ajax settings that will be able to support tags (the upcoming
feature). The goal here was to give more space to the writing area and
less to the settings. It makes it easier to write. Nuff said