Fixed a flex problem when autosaving.
What's happening is that flex is not recalculated when the autosave
message is displayed. That means that the text area is shrunk and
glitch. This seems to only happen when dealing with textarea as the
flexed items.

I fixed this by wrapping all the textareas in a <section> container.
That container is flexed and positioned relatively. Textareas are
absolute positionned in it.