Re-added S3 support and automatic upload
Images used to have their own section but this was replaced with a more
inline alternative. When creating an <img> tag in the textarea,
Ecrire will replace it if it has no src="" attribute with a form. That
form allow you to either click on it and select a file via the popup or
drag & drop an image. Once it's uploaded, Ecrire will replace the form
and them empty <img> tag with the image

This fix #52