Fixed the way Ecrire handles migration.
Whenever a user tries to access the console or start a server (in any
environment), Ecrire does a sanity check whether the database is fully
migrated. If it's not, Ecrire will ask if the user wants to migrate his
database (suggesting a backup). If the user accepts, the migration is
made and the server/console is fired up with the latest changes.

I did want to migrate without asking the user but the database changes
that a migration would bring is final. That means that if there's ever a
bug in a migration (but the migration is successful), it could mangle
the database. I rather have the confirmation from the user...