First attempt at adding labels to Ecrire.
This is in relation to #47 and it's almost complete. The tests are
missing, however.

I have chosen to use 'label' as the name for the tag system in Ecrire.
I believe this is more relevant.

As for what this commit brings to the table, here's the low down:

First, it brings a new two new models to the database: Label and
Property (The migration name is a little bit fucked but I will merge all
those migration before 1.0 so there will only be 1 migration per model
to start with fresh migrations).

Property is a HABTM model with polymorphic attributes, you can attach
any type of object to a Post. I'm not a huge find of this method however
as this brings a lot of complexity for very little gain. I always wanted
to play with the JSON type in (implemented in Postgresql 9.2) and I believe this would be
ideal. The controller however is solid so I think it's a good step
forward. The controller let you create and destroy new property on a
post (like a key-value store).

Now, you can only create label via this controller. Actually, more I'm
writing about this, more I'm thinking I should straight up start using
JSON type because it's fucking awesome and HABTM fucking sucks.

So long, and thanks for all the fish.