This is the essence of the new Hstore properties
The decorators, views, models and controllers needed to be refitted to
be able to handle the way Hstore do its thing.

One thing that is awesome with this change is that Hstore will be able
to store arbitrary data (Theme customization & storage, yay!)

First version of this is you can register a new class to a new property.
Then, you need to have that property bound to a model (usually under the
Property namespace)

When the controller will receive a POST or a DELETE, it will look at the
requested property and instantiate that class. It will then call
[create|destroy] depending on the requested action.