This commit wraps everything that touch to post properties
I really hate where I placed the header image but it's the best place I
could find for now. Now, the label works as well as the upload of image
as a header. The goal is that you can specify an image to be set as the
header. A template could then use that image to show on the post index
and in the header of that post.

It's still rough, but hey sometimes, committing is the right way to go