Fixed Warden not being able to authenticate.
Issue was two folds: First, the manager was not set to use any
strategies, even if there was a strategy available. So this first part
fixes that.

The second issue was a way more subtle. I was using @controller.env
variable which isn't the same as @request.env. Because of that, I didn't
have a working Request object that I could get information from to
authenticate the user.