Fixed the test coverage.
Basically, the previous test coverage needed to be called in a certain
way to work. It was retarded. So I decided I should fix that. Now, it
does some hacks using forking process to test different outcome since
many things that Ecrire does happens at boot time and it is bound to the
gobal namespace.