Implemented unordered list in the editor
Also changed the way the Editor handles string representation.
textContent is set through the browser and is pretty dumb. For list
parsing, I needed to split text content using \n but it was very hard to
get somewhere with textContent. So instead, I've decided to use
something that really sucks with JS and upgrade it through extensions:

`toString()`

This way, I can give different behavior for toString() depending on the
Element generated. This is very simple yet very effective. Also, since
toString() should do an object representation, it makes sense to use
this method.

I'm sure 100000 people will disagree with what I've done, but it works
for now and with very few lines of code, I managed to build the list
parser.