Re-implemented the UL and OL as well as cursor positioning.
I've decided to move more responsabilities to HTMLElement prototype
because I needed to do so calculation depending on the type of element
it was. I could have implemented a method that checks that, but it would
have been cumbersome. While I agree that it can bite me in the ass
eventually (It's global), I believe this is a good approach for now.