Re-implemented the save button.
Decided to squeeze the save button on the right of the content because I
want it to scroll as the use type so it's accessible. I also added a
small label to show when it was last saved.

Right now, it says "Update" but I think I'll revert it to "Save" because
it makes more sense in general and I don't want to be confusing people.

I also added moment.js because JS date is retarded and I wanted
something meaningful as the label for how long the post was saved.