Fixed white space problems with line feed.
The problem exists because CSS and content-editable are ennemies. Or I
don't know. For some stupid reason, if white-space is set to `normal`, it
will eat line space created by the user.

However, if I set it to `pre`, it will not eat the line but break words
at any letter which is retarded for *any* situation. I mean, why would
anyone wants to wrap a word anywhere?

So basically, I need to set white-space to `pre-wrap` which will
correctly keep the spaces while not breaking words.

Computers..