Implemented ctrl-s in the editor.
This is two fold: First, I stop propagation & save it like it would if
the user clicked on "save".

I also added a nudge animation that will trigger if the users ctrl-save
and there is nothing changed in the contnet of the post so the user
knows the save didn't go through.