Batch mutations and parsing is optimized
This is done by looking at the top most line that was affected by the
change and start the process from there. This change has knock on effect
that fixed few issues with cursor positioning. A few checks in
appended(), updated() and removed() might become unecessary which might
bring a whole lot of speed improvement.