Can now create new tags and also reorganized the code base
With this commit, #114 is now feature complete. While I think it still
is a little rough around the edges, I think it doesn't matter for now as
I've been waiting to use this feature for quite a while now ;)

Moreover, the lack of polish doesn't make the feature *unusable* so
there shouldn't be a huge difference in term of usability from this
version to the more polish one that will come along eventually.