Parsers were just added as they were eval'ed by the browser
This led to a problem when some parsers needs to parse the content of
the text before others. Due to this shortcoming, I changed the Parser's
container from Array to a Class with it's internal store and a sort()
method which returns the list of parsers in the preferred order.